refactor updateRelease
