+++ src/publish-release.ts
@@ -1,17 +1,22 @@
-import * as http from "@actions/http-client";
+import * as core from "@actions/core";
+import * as github from "./github-mini";
 
-export type MakeLatest = "true" | "false" | "legacy";
+const handleGitHubError = (msg: string, error: github.GitHubError): never => {
+  core.error(
+    `${msg}: unexpected status code: ${error.statusCode}, error: ${JSON.stringify(error.error)}`,
+  );
+  throw new Error(`unexpected status code: ${error.statusCode}`);
+};
 
 // Options is the options for publish function.
 interface Options {
-  github_token: string;
+  client: github.Client;
+
   owner: string;
   repo: string;
   id: string;
   discussion_category_name: string;
-  make_latest?: MakeLatest;
-
-  updateRelease?: (params: ReposUpdateReleaseParams) => Promise<void>;
+  make_latest?: github.MakeLatest;
 }
 
 // publish publishes the release.
@@ -22,9 +27,8 @@ export async function publish(opt: Options): Promise<void> {
   const discussion_category_name =
     opt.discussion_category_name !== "" ? opt.discussion_category_name : undefined;
   const make_latest = opt.make_latest;
-  const updater = opt.updateRelease || updateRelease;
-  await updater({
-    github_token: opt.github_token,
+
+  const resp = await opt.client.updateRelease({
     owner,
     repo,
     id: opt.id,
@@ -32,49 +36,7 @@ export async function publish(opt: Options): Promise<void> {
     discussion_category_name,
     make_latest,
   });
+  if (resp.isFailure()) {
+    return handleGitHubError("failed to publish new release", resp.value);
+  }
 }
-
-const newGitHubClient = (token: string): http.HttpClient => {
-  return new http.HttpClient("shogo82148-actions-create-release/v1", [], {
-    headers: {
-      Authorization: `Bearer ${token}`,
-      Accept: "application/vnd.github+json",
-      "X-GitHub-Api-Version": "2022-11-28",
-    },
-  });
-};
-
-interface ReposUpdateReleaseParams {
-  github_token: string;
-  owner: string;
-  repo: string;
-  id: string;
-  draft: boolean;
-  discussion_category_name?: string | undefined;
-  make_latest?: MakeLatest | undefined;
-}
-
-// minium implementation of update a release API
-// https://docs.github.com/en/rest/releases/releases?apiVersion=2022-11-28#update-a-release
-const updateRelease = async (params: ReposUpdateReleaseParams): Promise<void> => {
-  const client = newGitHubClient(params.github_token);
-  const raw: { [key: string]: string | boolean } = {
-    draft: params.draft,
-  };
-  if (params.discussion_category_name) {
-    raw["discussion_category_name"] = params.discussion_category_name;
-  }
-  if (params.make_latest) {
-    raw["make_latest"] = params.make_latest;
-  }
-
-  const body = JSON.stringify(raw);
-  const api = process.env["GITHUB_API_URL"] || "https://api.github.com";
-  const url = `${api}/repos/${params.owner}/${params.repo}/releases/${params.id}`;
-  const resp = await client.request("PATCH", url, body, {});
-  const statusCode = resp.message.statusCode;
-  const contents = await resp.readBody();
-  if (statusCode !== 200) {
-    throw new Error(`unexpected status code: ${statusCode}\n${contents}`);
-  }
-};
